Further changes to support the Optima file browser
Added a background to TGUI_ScrollPane

Set the height of TGUI_TextField in the constructor, rather than draw, as
it won't change, and this allows it to be used in sizing calculations.

Fixed a bug in TGUI_Frame.getAbsoluteChildPosition() which didn't
properly recurse down, so failed if the child widget was a splitter.

Made items in TGUI_List clickable, so to support used in the file browser.